requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-jose==3.3.0
httpx==0.27.0
pydantic-settings==2.2.1
